refactor: 修正遊戲切換條件
